fix: Compilation Error, DMG building
I think this is the final fix!

There was an incompatible return type error. I presume this is
because since this was last edited, the Obj-C compiler has gotten
more strict.

Secondly, it was failing to build the DMG, becuase of permissioning
errors. Turning off sandboxing fixed it. Not a great option but sure
look, it builds.